requirements.txt
streamlit
pandas
plotly
binance-futures-connector